chakrashim: fix internal promise behavior
`test-promise-internal-creation` was failing because of the `state` and
`result` properties we were attaching to it.  Moved those fields into
the external object.

This change also enables more promise tests.

PR-URL: https://github.com/nodejs/node-chakracore/pull/471
