+++ test/parallel/parallel.status
@@ -34,7 +34,6 @@ test-async-hooks-promise-triggerid : SKIP
 test-async-wrap-promise-after-enabled : SKIP
 test-buffer-bindingobj-no-zerofill : SKIP
 test-buffer-sharedarraybuffer : SKIP
-test-child-process-promisified : SKIP
 test-crypto-dh : SKIP
 test-crypto-dh-leak : SKIP
 test-domain-abort-on-uncaught : SKIP
@@ -49,11 +48,9 @@ test-domain-no-error-handler-abort-on-uncaught-7 : SKIP
 test-domain-no-error-handler-abort-on-uncaught-8 : SKIP
 test-domain-no-error-handler-abort-on-uncaught-9 : SKIP
 test-domain-promise : SKIP
-test-fs-promisified : SKIP
 test-fs-stat : SKIP
 test-http-pipeline-flood : SKIP
 test-http-same-map : SKIP
-test-http2-client-promisify-connect : SKIP
 test-http2-server-rst-stream : SKIP
 test-inspect-async-hook-setup-at-inspect : SKIP
 test-intl : SKIP
@@ -62,7 +59,6 @@ test-module-main-extension-lookup : SKIP
 test-performance-function : SKIP
 test-performance-gc : SKIP
 test-process-env-symbols : SKIP
-test-promise-internal-creation : SKIP
 test-promises-unhandled-symbol-rejections : SKIP
 test-regress-GH-12371 : SKIP
 test-repl : SKIP
@@ -73,7 +69,6 @@ test-repl-pretty-stack : SKIP
 test-repl-sigint : SKIP
 test-repl-tab-complete : SKIP
 test-string-decoder : SKIP
-test-timers-promisified : SKIP
 
 # ChakraCore does not support the chrome tracing framework
 test-trace-events-all : PASS,SKIP
@@ -89,7 +84,6 @@ test-util : SKIP
 test-util-format-shared-arraybuffer : SKIP
 test-util-inspect-proxy : SKIP
 test-util-inspect : SKIP
-test-util-promisify : SKIP
 test-v8-serdes : SKIP
 test-v8-serdes-sharedarraybuffer : SKIP
 test-vm-cached-data : SKIP
